Fix cmake complaining about apriltag::apriltag
Can't build on Ubuntu 20.04 because cmake 
keeps running into an error when trying to compile:

CMake Error at CMakeLists.txt:24 (add_library):
Target "AprilTagNode" links to target "apriltag::apriltag" but the target
was not found.  Perhaps a find_package() call is missing for an IMPORTED
target, or an ALIAS target is missing?